Refactored intro to its own component
