+++ src/App.tsx
@@ -1,6 +1,6 @@
 
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import IntroWrapper from './pages/Intro';
+import Intro from './pages/Intro';
 import Ambassador from './pages/Ambassador';
 import Home from './pages/Home';
 import Sponsors from './pages/Sponsors';
@@ -10,7 +10,7 @@ function App() {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<IntroWrapper />} />
+        <Route path="/" element={<Intro />} />
         <Route path="/home" element={<Home />} />
         <Route path="/sponsors" element={<Sponsors />} />
         <Route path="/gallery" element={<Gallery />} />
